XProcException refactoring
Working towards a clearer distinction between XProc errors and Java
exceptions, and cleaning up the code.

All constructors and static instantiation methods are now build upon
one main constructor. Arguments are named to better represent what
they are. Every constructor has a subset of the same four arguments:
code, location, message and cause.

Each argument can have one of several types. For example, the location
can be specified as a SourceLocator[], XdmNode, Step, XStep,
TransformerException or Throwable.

The message argument can now also be provided as a XdmNode instead of
a string. The node will be included in the serialized error.

A clear distinction is made between the cause of the XProc
error (getErrorCause) vs. the cause of the Java
exception (getCause). The latter is now provided through the message
argument. In order to provide a Throwable combined with a String
message, the Throwable and the message need to be wrapped in a new
Throwable first and then an XProcException can be created from it.

The javaError method, now called "fromException", was simplified and
instead the rebaseOnto method, now called "rebase" was made more
powerful.